requirements.txt
frida-tools==8.1.3